refactor(chat): isolate index file hardening
